+++ packages/core/helper-plugin/lib/src/components/CheckPermissions/CheckPermissions.js
@@ -1,9 +1,8 @@
 import { useEffect, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useNotification } from '../../hooks/useNotification';
-
-import hasPermissions from '../../utils/hasPermissions';
 import { useRBACProvider } from '../../hooks/useRBACProvider';
+import { hasPermissions } from '../../utils/hasPermissions';
 
 // NOTE: this component is very similar to the CheckPagePermissions
 // except that it does not handle redirections nor loading state
