+++ packages/core/helper-plugin/lib/src/components/ContentBox/ContentBox.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { Row } from '@strapi/parts/Row';
 import { Stack } from '@strapi/parts/Stack';
 import { Text } from '@strapi/parts/Text';
-import customEllipsis from '../../utils/customEllipsis';
+import { customEllipsis } from '../../utils/customEllipsis';
 
 const IconWrapper = styled(Row)`
   margin-right: ${({ theme }) => theme.spaces[6]};
