+++ packages/core/helper-plugin/lib/src/providers/NotificationsProvider/NotificationsProvider.js
@@ -1,0 +1,23 @@
+/**
+ *
+ * NotificationsProvider
+ *
+ */
+import React from 'react';
+import PropTypes from 'prop-types';
+import { NotificationsContext } from '../../contexts/NotificationsContext';
+
+const NotificationsProvider = ({ children, toggleNotification }) => {
+  return (
+    <NotificationsContext.Provider value={{ toggleNotification }}>
+      {children}
+    </NotificationsContext.Provider>
+  );
+};
+
+NotificationsProvider.propTypes = {
+  children: PropTypes.node.isRequired,
+  toggleNotification: PropTypes.func.isRequired,
+};
+
+export { NotificationsProvider };
