+++ packages/core/helper-plugin/lib/src/utils/await-to-js/index.js
@@ -1,14 +1,1 @@
-const to = (promise, errorExt) => {
-  return promise
-    .then(function(data) {
-      return [null, data];
-    })
-    .catch(function(err) {
-      if (errorExt) {
-        Object.assign(err, errorExt);
-      }
-      return [err, undefined];
-    });
-};
-
-export default to;
+export * from './to';
