+++ packages/core/helper-plugin/lib/src/utils/await-to-js/to.js
@@ -1,0 +1,14 @@
+const to = (promise, errorExt) => {
+  return promise
+    .then(function(data) {
+      return [null, data];
+    })
+    .catch(function(err) {
+      if (errorExt) {
+        Object.assign(err, errorExt);
+      }
+      return [err, undefined];
+    });
+};
+
+export { to };
