+++ packages/core/helper-plugin/lib/src/utils/customEllipsis/customEllipsis.js
@@ -1,0 +1,6 @@
+const customEllipsis = title => {
+  const endSubstring = title.indexOf('registrationToken') + 20;
+  return `${title.substring(0, endSubstring)}...${title.substring(title.length - 3, title.length)}`;
+};
+
+export { customEllipsis };
