+++ packages/core/helper-plugin/lib/src/utils/customEllipsis/index.js
@@ -1,6 +1,1 @@
-const customEllipsis = title => {
-  const endSubstring = title.indexOf('registrationToken') + 20;
-  return `${title.substring(0, endSubstring)}...${title.substring(title.length - 3, title.length)}`;
-};
-
-export default customEllipsis;
+export * from './customEllipsis';
