+++ packages/core/helper-plugin/lib/src/utils/getFileExtension/getFileExtension.js
@@ -1,3 +1,3 @@
 const getFileExtension = ext => (ext ? ext.substr(1) : null);
 
-export default getFileExtension;
+export { getFileExtension };
