+++ packages/core/helper-plugin/lib/src/utils/getFileExtension/index.js
@@ -1,0 +1,1 @@
+export * from './getFileExtension';
