+++ packages/core/helper-plugin/lib/src/utils/getFileExtension/tests/getFileExtension.test.js
@@ -1,4 +1,4 @@
-import getFileExtension from '../getFileExtension';
+import { getFileExtension } from '../getFileExtension';
 
 describe('HELPER PLUGIN | utils | getFileExtension', () => {
   it('should return undefined if ext does not exits', () => {
