+++ packages/core/helper-plugin/lib/src/utils/hasPermissions/tests/hasPermissions.test.js
@@ -1,4 +1,5 @@
-import hasPermissions, {
+import {
+  hasPermissions,
   findMatchingPermissions,
   formatPermissionsForRequest,
   shouldCheckPermissions,
