+++ packages/core/helper-plugin/lib/src/utils/prefixFileUrlWithBackendUrl/index.js
@@ -1,0 +1,1 @@
+export * from './prefixFileUrlWithBackendUrl';
