+++ packages/core/helper-plugin/lib/src/utils/prefixFileUrlWithBackendUrl/prefixFileUrlWithBackendUrl.js
@@ -2,4 +2,4 @@ const prefixFileUrlWithBackendUrl = fileURL => {
   return !!fileURL && fileURL.startsWith('/') ? `${strapi.backendURL}${fileURL}` : fileURL;
 };
 
-export default prefixFileUrlWithBackendUrl;
+export { prefixFileUrlWithBackendUrl };
