+++ packages/core/helper-plugin/lib/src/utils/prefixPluginTranslations/index.js
@@ -1,9 +1,1 @@
-const prefixPluginTranslations = (trad, pluginId) => {
-  return Object.keys(trad).reduce((acc, current) => {
-    acc[`${pluginId}.${current}`] = trad[current];
-
-    return acc;
-  }, {});
-};
-
-export default prefixPluginTranslations;
+export * from './prefixPluginTranslations';
