+++ packages/core/helper-plugin/lib/src/utils/prefixPluginTranslations/prefixPluginTranslations.js
@@ -1,0 +1,9 @@
+const prefixPluginTranslations = (trad, pluginId) => {
+  return Object.keys(trad).reduce((acc, current) => {
+    acc[`${pluginId}.${current}`] = trad[current];
+
+    return acc;
+  }, {});
+};
+
+export { prefixPluginTranslations };
