+++ packages/core/helper-plugin/lib/src/utils/pxToRem/index.js
@@ -1,3 +1,1 @@
-const pxToRem = px => `${px / 16}rem`;
-
-export default pxToRem;
+export * from './pxToRem';
