+++ packages/core/helper-plugin/lib/src/utils/pxToRem/pxToRem.js
@@ -1,0 +1,3 @@
+const pxToRem = px => `${px / 16}rem`;
+
+export { pxToRem };
