+++ packages/core/helper-plugin/lib/src/utils/setHexOpacity/index.js
@@ -1,6 +1,1 @@
-const setHexOpacity = (hex, alpha) =>
-  `${hex}${Math.floor(alpha * 255)
-    .toString(16)
-    .padStart(2, 0)}`;
-
-export default setHexOpacity;
+export * from './setHexOpacity';
