+++ packages/core/helper-plugin/lib/src/utils/setHexOpacity/setHexOpacity.js
@@ -1,0 +1,6 @@
+const setHexOpacity = (hex, alpha) =>
+  `${hex}${Math.floor(alpha * 255)
+    .toString(16)
+    .padStart(2, 0)}`;
+
+export { setHexOpacity };
